requirements.txt
python-telegram-bot==13.12
beautifulsoup4==4.11.1
requests==2.28.0
lxml==4.9.0
PyGithub==1.55